fix : 상세페이지 응답값 오류사항 수정

+++ src/main/java/com/sparta/village/domain/reservation/service/ReservationService.java
@@ -1,5 +1,6 @@
 package com.sparta.village.domain.reservation.service;
 
+import com.sparta.village.domain.product.entity.Product;
 import com.sparta.village.domain.product.repository.ProductRepository;
 import com.sparta.village.domain.product.service.ProductService;
 import com.sparta.village.domain.reservation.dto.AcceptReservationResponseDto;
@@ -10,6 +11,7 @@ import com.sparta.village.domain.reservation.dto.ReservationResponseDto;
 import com.sparta.village.domain.reservation.dto.StatusRequestDto;
 import com.sparta.village.domain.reservation.entity.Reservation;
 import com.sparta.village.domain.reservation.repository.ReservationRepository;
+import com.sparta.village.domain.user.entity.User;
 import com.sparta.village.domain.user.service.KakaoUserService;
 import com.sparta.village.global.exception.CustomException;
 import com.sparta.village.global.exception.ErrorCode;
@@ -31,63 +33,57 @@ public class ReservationService {
 
 
     @Transactional
-    public ResponseEntity<ResponseMessage> reserve(Long productId, ReservationRequestDto requestDto, Long userId) {
+    public ResponseEntity<ResponseMessage> reserve(Long productId, ReservationRequestDto requestDto, User user) {
         //제품 있는지 체크(제품 등록 코드 완성되면 추가하기!!)
 
-        if (!productRepository.existsById(productId)) {
-            throw new CustomException(ErrorCode.PRODUCT_NOT_FOUND);
-        }
+        Product product = productRepository.findById(productId).orElseThrow(
+                () -> new CustomException(ErrorCode.PRODUCT_NOT_FOUND)
+        );
 
         //예약 가능한 날짜인지 체크
-        if (reservationRepository.checkOverlapDateByProductId(productId, requestDto.getStartDate(), requestDto.getEndDate())) {
+        if (reservationRepository.checkOverlapDateByProduct(product, requestDto.getStartDate(), requestDto.getEndDate())) {
             throw new CustomException(ErrorCode.DUPLICATE_RESERVATION_DATE);
         }
         //예약 테이블에 저장
-        reservationRepository.saveAndFlush(new Reservation(productId, userId, requestDto));
+        reservationRepository.saveAndFlush(new Reservation(product, user, requestDto));
         return ResponseMessage.SuccessResponse("예약 되었습니다.", "");
     }
 
     @Transactional
-    public ResponseEntity<ResponseMessage> deleteReservation(Long id, Long userId) {
-        checkReservationId(id);
-        checkReservationUserId(userId);
+    public ResponseEntity<ResponseMessage> deleteReservation(Long id, User user) {
+        Reservation reservation = findReservationById(id);
+        if (!reservation.getUser().getId().equals(user.getId())) {
+            throw new CustomException(ErrorCode.NOT_AUTHOR);
+        }
         reservationRepository.deleteById(id);
         return ResponseMessage.SuccessResponse("예약 취소되었습니다.", "");
     }
 
     @Transactional
     public ResponseEntity<ResponseMessage> changeStatus(Long id, StatusRequestDto requestDto, Long userId) {
-        checkReservationId(id);
+        Reservation reservation = findReservationById(id);
 //        productService.checkProductOwner(reservationRepository.findProductIdById(id), userId);
-
-        reservationRepository.updateStatus(id, requestDto.getStatus());
+        reservationRepository.updateStatus(reservation.getId(), requestDto.getStatus());
         return ResponseMessage.SuccessResponse("상태 변경되었습니다.", "");
     }
 
 
-    private void checkReservationId(Long id) {
-        if (!reservationRepository.existsById(id)) {
-            throw new CustomException(ErrorCode.RESERVATION_NOT_FOUND);
-        }
+    private Reservation findReservationById(Long id) {
+        return reservationRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.RESERVATION_NOT_FOUND));
     }
 
 
-    private void checkReservationUserId(Long userId) {
-        if (!reservationRepository.existsByUserId(userId)) {
-            throw new CustomException(ErrorCode.NOT_AUTHOR);
-        }
-    }
-
-    public List<ReservationResponseDto> getReservationListByProductId(Long id) {
-        List<ReservationResponseDto> reservationList = reservationRepository.findAllReservationDtoByProductId(id);
-        reservationList.forEach(r -> r.setNickname(userService.getUserByUserId(r.getNickname()).getNickname()));
-        return reservationList;
+    public List<ReservationResponseDto> getReservationList(Long id){
+        return reservationRepository.findByProductId(id).stream()
+                .map(r -> new ReservationResponseDto(r.getId(), r.getStartDate(), r.getEndDate(), r.getStatus(), r.getUser().getNickname(), r.getProduct().getId())).toList();
     }
 
     public ResponseEntity<ResponseMessage> getAcceptedReservationList() {
-        List<AcceptReservationResponseDto> acceptReservationList = reservationRepository.findAcceptedReservationDto();
-        acceptReservationList.forEach(r -> r.setOwnerNickname(userService.getUserByUserId(r.getOwnerNickname()).getNickname()));
-        acceptReservationList.forEach(r -> r.setCustomerNickname(userService.getUserByUserId(r.getCustomerNickname()).getNickname()));
+        List<AcceptReservationResponseDto> acceptReservationList = reservationRepository.findByStatus("accepted").stream()
+                .map(r -> new AcceptReservationResponseDto(r.getId(), r.getUser().getNickname(), r.getProduct().getUser().getNickname())).toList();
+//        List<AcceptReservationResponseDto> acceptReservationList = reservationRepository.findAcceptedReservationDto();
+//        acceptReservationList.forEach(r -> r.setOwnerNickname(userService.getUserByUserId(r.getOwnerNickname()).getNickname()));
+//        acceptReservationList.forEach(r -> r.setCustomerNickname(userService.getUserByUserId(r.getCustomerNickname()).getNickname()));
         return ResponseMessage.SuccessResponse("검색완료 되었습니다.", acceptReservationList);
     }
 }
